requirements.txt
Django==1.6.2
South==0.8.4
django-extensions==1.3.3
django-floppyforms==1.1
django-debug-toolbar==1.0.1
django-extra-views==0.6.3

# allauth with trnslation to hebrew, requires git installed :-)
# django-allauth==0.13.0
-e git+https://github.com/nonZero/django-allauth.git#egg=django_allauth-dev

bleach==1.2.2
Markdown==2.3.1
Pygments==1.6

ipython==1.1.0
python-dateutil==2.1
